fix(lint): allow unexpected cfgs
This commit add an exception for the lint `unexpected_cfgs` in Cargo.toml

This rule does not get triggered with the default code but will be
triggered as soon as the user adds a `pins.rs` file.

This is only to avoid confusion until a better solution is found for
assigning pins.

Signed-off-by: Nils Ponsard <nils.ponsard@inria.fr>